+++ CHANGELOG.md
@@ -15,6 +15,11 @@ All notable changes to this project will be documented in this file.
   is not corrupted or tampered with.
 - Added `criticalup archive` which creates an archive of the toolchain for cold storage or backup.
 
+## Fixed
+
+- Bug when using `--offline` mode to install with expired revocation info ends in installation failure. To
+  support proper `--offline` mode, the expiration date on revocation info hash must be ignored.
+
 ## [1.1.0] - 2024-08-28
 
 ## Added